Add assignment endpoint
