+++ client/src/typings/result.typings.ts
@@ -4,11 +4,3 @@ export interface Result {
   username: string;
   enteredGame: Signup;
 }
-
-export interface PostPlayerAssignmentResponse {
-  message: string;
-  resultMessage: string;
-  results: Result[];
-  startTime: string;
-  status: 'success';
-}
