+++ shared/typings/api/assignment.ts
@@ -1,0 +1,9 @@
+import { Result } from 'client/typings/result.typings';
+
+export interface PostPlayerAssignmentResponse {
+  message: string;
+  resultMessage: string;
+  results: readonly Result[];
+  startTime: string;
+  status: 'success';
+}
